Revert "Feature login successful"
