--- src/app/pages/login/login.module.ts
@@ -1,19 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { IonicModule } from '@ionic/angular';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { LoginPage } from './login.page';
-import { LoginPageRoutingModule } from './login-routing.module';
-
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-    LoginPageRoutingModule,
-    ReactiveFormsModule
-  ],
-  declarations: [LoginPage]
-})
-export class LoginPageModule {}
