--- src/app/pages/login/login.page.html
@@ -1,34 +1,0 @@
-<ion-header [translucent]="true">
-    <ion-toolbar>
-      <ion-title>
-        Login
-      </ion-title>
-    </ion-toolbar>
-  </ion-header>
-
-  <ion-content [fullscreen]="true">
-
-      <ion-card>
-        <form class="form" [formGroup]="validations_form" (ngSubmit)="tryLogin(validations_form.value)" >
-        <ion-item>
-          <ion-label position="floating" color="primary" >Email</ion-label>
-          <ion-input type="text" formControlName="email"></ion-input>
-        </ion-item>
-        <ion-item>
-          <ion-label position="floating" color="primary">Password</ion-label>
-          <ion-input type="password" formControlName="password" class="form-control" required ></ion-input>
-        </ion-item>
-        <ion-item>
-          <ion-button color="success" type="submit">
-            Login
-          </ion-button>
-        </ion-item>
-      </form>
-        <ion-item>
-          <ion-button color="light" (click)="gotoSignup()" >
-            Signup
-          </ion-button>
-        </ion-item>
-    </ion-card>
-
-  </ion-content>
