--- src/app/pages/login/login.page.ts
@@ -1,65 +1,0 @@
-import { AuthenticationService } from './../../services/authentication.service';
-import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
-import { NavController } from '@ionic/angular';
-
-@Component({
-  selector: 'app-login',
-  templateUrl: 'login.page.html',
-  styleUrls: ['login.page.scss'],
-})
-export class LoginPage implements OnInit {
-
-  validations_form: FormGroup;
-  errorMessage: string = '';
-  successMessage: string = '';
-
-  validation_messages = {
-    'email': [
-      { type: 'required', message: 'Email is required.' },
-      { type: 'pattern', message: 'Enter a valid email' }
-    ],
-    'password': [
-      { type: 'required', message: 'Password is required' },
-      { type: 'minlength', message: 'Password must be at least 6 characters long' }
-    ]
-  };
-
-  constructor(
-    private NavCtrl: NavController,
-    private authService: AuthenticationService,
-    private formBuilder: FormBuilder
-  ) {}
-
-  ngOnInit() {
-    this.validations_form = this.formBuilder.group({
-      email: new FormControl('', Validators.compose([
-        Validators.required,
-        Validators.pattern('^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+.[a-zA-Z0-9-.]+$')
-      ])),
-      password: new FormControl('', Validators.compose([
-        Validators.required,
-        Validators.minLength(6)
-      ]))
-    })
-  }
-
-  tryLogin(value) {
-    this.authService.loginUser(value)
-    .then( res => {
-      console.log(res);
-      this.errorMessage = "";
-      this.successMessage = "You have successfully logged in.";
-      this.NavCtrl.navigateForward('/home');
-    }, err => {
-      console.log(err);
-      this.errorMessage = err.message;
-      this.successMessage = "";
-    })
-  }
-
-  gotoSignup() {
-    this.NavCtrl.navigateForward('/signup');
-  }
-
-}
