--- src/app/pages/signup/signup.module.ts
@@ -1,20 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { IonicModule } from '@ionic/angular';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { SignupPage } from './signup.page';
-
-import { SignupPageRoutingModule } from './signup-routing.module';
-
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-    SignupPageRoutingModule,
-    ReactiveFormsModule
-  ],
-  declarations: [SignupPage]
-})
-export class SignupPageModule {}
