--- src/app/pages/signup/signup.page.html
@@ -1,39 +1,0 @@
-<ion-header [translucent]="true">
-  <ion-toolbar>
-    <ion-title>
-      Signup
-    </ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content [fullscreen]="true">
-  <!-- <ion-header collapse="condense">
-    <ion-toolbar>
-      <ion-title size="large">Signup Page</ion-title>
-    </ion-toolbar>
-  </ion-header> -->
-<form class="form" [formGroup]="validations_form" (ngSubmit)="trySignup(validations_form.value)" >
-  <!-- <div id="container"> -->
-    <ion-card>
-      <ion-item>
-        <ion-label position="floating" color="primary" >Email</ion-label>
-        <ion-input type="text" formControlName="email" ></ion-input>
-      </ion-item>
-      <ion-item>
-        <ion-label position="floating" color="primary">Password</ion-label>
-        <ion-input type="password" formControlName="password" class="form-control" required ></ion-input>
-      </ion-item>
-      <ion-item>
-        <ion-button color="success" type="submit" (ngSubmit)="trySignup(value)" >
-          Signup
-        </ion-button>
-      </ion-item>
-      <ion-item>
-        <ion-button color="light" [routerLink]="['/login']" >
-          Login
-        </ion-button>
-      </ion-item>
-  </ion-card>
-  <!-- </div> -->
-</form>
-</ion-content>
