--- src/app/pages/signup/signup.page.ts
@@ -1,64 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-import { NavController } from '@ionic/angular';
-import { FormGroup, FormBuilder, Validators, FormControl} from '@angular/forms';
-import { AuthenticationService } from '../../services/authentication.service';
-
-@Component({
-  selector: 'app-signup',
-  templateUrl: 'signup.page.html',
-  styleUrls: ['signup.page.scss'],
-})
-export class SignupPage implements OnInit {
-
-  validations_form: FormGroup;
-  errorMessage: string = '';
-  successMessage: string = '';
-
-  validation_messages = {
-    'email': [
-      { type: 'required', message: 'Email is required.' },
-      { type: 'pattern', message: 'Enter a valid email' }
-    ],
-    'password': [
-      { type: 'required', message: 'Password is required' },
-      { type: 'minlength', message: 'Password must be at least 6 characters long' }
-    ]
-  };
-
-  constructor(
-    private NavCtrl : NavController,
-    private authService: AuthenticationService,
-    private formBuilder: FormBuilder
-  ) {}
-
-  ngOnInit() {
-    this.validations_form = this.formBuilder.group({
-      email: new FormControl('', Validators.compose([
-        Validators.required,
-        Validators.pattern('^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+.[a-zA-Z0-9-.]+$')
-      ])),
-      password: new FormControl('', Validators.compose([
-        Validators.required,
-        Validators.minLength(6)
-      ]))
-    })
-  }
-
-  trySignup(value) {
-    this.authService.signupUser(value)
-    .then( res => {
-      console.log(res);
-      this.errorMessage="";
-      this.successMessage="Your account has been created.";
-    },err => {
-      console.log(err);
-      this.errorMessage=err.message;
-      this.successMessage="";
-    })
-  }
-
-  gotoLogin(){
-    this.NavCtrl.navigateBack('');
-  }
-
-}
