--- src/app/services/authentication.service.ts
@@ -1,50 +1,0 @@
-import { Injectable } from '@angular/core';
-import { AngularFireAuth} from '@angular/fire/auth';
-// import { rejects } from 'assert';
-// import { resolve } from 'dns';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class AuthenticationService {
-
-  constructor(
-    private afAuth: AngularFireAuth
-  ) { }
-
-  signupUser(value){
-    return new Promise<any>((resolve,reject) => {
-      this.afAuth.createUserWithEmailAndPassword(value.email, value.password).then(res => resolve(res), err=> reject(err))
-    })
-  }
-
-  loginUser(value){
-    return new Promise<any>((resolve, reject) => {
-      this.afAuth.signInWithEmailAndPassword(value.email, value.password).then(
-        res => resolve(res),
-        err => reject(err)
-      )
-    })
-  }
-
-  logoutUser(){
-    return new Promise((resolve, reject) => {
-      if(this.afAuth.currentUser){
-        this.afAuth.signOut().then(()=> {
-          console.log("LOG OUT");
-          resolve();
-        }).catch((error) => {
-          reject();
-        })
-      }
-    })
-  }
-
-  userDetails(){
-    return this.afAuth.user
-  }
-
-  userList(){
-
-  }
-}
